➕add: api/audioHistory.ts creation
